# app/shared/ml_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ADSMOD Model Fitting Backend ML Service",
    "description": "ADSMOD machine learning service for dataset preparation, training, checkpoint management, and model workflows.",
    "version": "1.2.0"
  },
  "paths": {
...
    "/api/training/dataset-source": {
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Delete Dataset Source",
        "operationId": "delete_dataset_source_api_training_dataset_source_delete",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(nist|uploaded)$",
              "title": "Source"
            }
          },
          {
            "name": "dataset_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9_. -]+$",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSourceDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/build-dataset": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Build Training Dataset",
        "operationId": "build_training_dataset_api_training_build_dataset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training/dataset-info": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Dataset Info",
        "operationId": "get_dataset_info_api_training_dataset_info_get",
        "parameters": [
          {
            "name": "dataset_label",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]{0,63}$",
              "default": "default",
              "title": "Dataset Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/dataset": {
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Clear Training Dataset",
        "operationId": "clear_training_dataset_api_training_dataset_delete",
        "parameters": [
          {
            "name": "dataset_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64,
                  "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]{0,63}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dataset Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training/jobs/{job_id}": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Dataset Job Status",
        "operationId": "get_dataset_job_status_api_training_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Cancel Dataset Job",
        "operationId": "cancel_dataset_job_api_training_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training/checkpoints/{checkpoint_name}": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Checkpoint Details",
        "operationId": "get_checkpoint_details_api_training_checkpoints__checkpoint_name__get",
        "parameters": [
          {
            "name": "checkpoint_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Checkpoint Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointFullDetailsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Delete Checkpoint",
        "operationId": "delete_checkpoint_api_training_checkpoints__checkpoint_name__delete",
        "parameters": [
          {
            "name": "checkpoint_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$",
              "title": "Checkpoint Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/start": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Start Training",
        "operationId": "start_training_api_training_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/resume": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Resume Training",
        "operationId": "resume_training_api_training_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeTrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetBuildRequest": {
        "properties": {
          "sample_size": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.01,
            "title": "Sample Size",
            "default": 1.0
          },
          "validation_size": {
            "type": "number",
            "maximum": 0.5,
            "minimum": 0.05,
            "title": "Validation Size",
            "default": 0.2
          },
          "min_measurements": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Min Measurements",
            "default": 1
          },
          "max_measurements": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 5.0,
            "title": "Max Measurements",
            "default": 30
          },
          "smile_sequence_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 5.0,
            "title": "Smile Sequence Size",
            "default": 20
          },
          "max_pressure": {
            "type": "number",
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Max Pressure",
            "default": 10000.0
          },
          "max_uptake": {
            "type": "number",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Max Uptake",
            "default": 20.0
          },
          "reference_checkpoint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Checkpoint"
          },
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetSelection"
            },
            "type": "array",
            "minItems": 1,
            "title": "Datasets"
          },
          "dataset_label": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]{0,63}$",
            "title": "Dataset Label",
            "default": "default"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DatasetBuildRequest"
      },
      "DatasetInfoResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "dataset_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9 _-]{0,63}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Label"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "sample_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Size"
          },
          "validation_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Size"
          },
          "min_measurements": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Measurements"
          },
          "max_measurements": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Measurements"
          },
          "smile_sequence_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smile Sequence Size"
          },
          "max_pressure": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pressure"
          },
          "max_uptake": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Uptake"
          },
          "total_samples": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Samples"
          },
          "train_samples": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Train Samples"
          },
          "validation_samples": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Samples"
          },
          "smile_vocabulary_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smile Vocabulary Size"
          },
          "adsorbent_vocabulary_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adsorbent Vocabulary Size"
          },
          "normalization_stats": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalization Stats"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "DatasetInfoResponse"
      },
...
      "DatasetSourceDeleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "DatasetSourceDeleteResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStartResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "poll_interval": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poll Interval"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_type",
          "status",
          "message"
        ],
        "title": "JobStartResponse"
      },
...
      "OperationStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "OperationStatusResponse"
      },
...
    }
  }
}